removed directoryrefresher.h from organizercore.h
