+++ src/shared/directoryentry.cpp
@@ -73,105 +73,6 @@ static bool DirCompareByName(const DirectoryEntry *lhs, const DirectoryEntry *rh
 }
 
 
-DirectoryStats::DirectoryStats()
-{
-  std::memset(this, 0, sizeof(DirectoryStats));
-}
-
-DirectoryStats& DirectoryStats::operator+=(const DirectoryStats& o)
-{
-  dirTimes += o.dirTimes;
-  fileTimes += o.fileTimes;
-  sortTimes += o.sortTimes;
-
-  subdirLookupTimes += o.subdirLookupTimes;
-  addDirectoryTimes += o.addDirectoryTimes;
-
-  filesLookupTimes += o.filesLookupTimes;
-  addFileTimes += o.addFileTimes;
-  addOriginToFileTimes += o.addOriginToFileTimes;
-  addFileToOriginTimes += o.addFileToOriginTimes;
-  addFileToRegisterTimes += o.addFileToRegisterTimes;
-
-  originExists += o.originExists;
-  originCreate += o.originCreate;
-  originsNeededEnabled += o.originsNeededEnabled;
-
-  subdirExists += o.subdirExists;
-  subdirCreate += o.subdirCreate;
-
-  fileExists += o.fileExists;
-  fileCreate += o.fileCreate;
-  filesInsertedInRegister += o.filesInsertedInRegister;
-  filesAssignedInRegister += o.filesAssignedInRegister;
-
-  return *this;
-}
-
-std::string DirectoryStats::csvHeader()
-{
-  QStringList sl = {
-    "dirTimes",
-    "fileTimes",
-    "sortTimes",
-    "subdirLookupTimes",
-    "addDirectoryTimes",
-    "filesLookupTimes",
-    "addFileTimes",
-    "addOriginToFileTimes",
-    "addFileToOriginTimes",
-    "addFileToRegisterTimes",
-    "originExists",
-    "originCreate",
-    "originsNeededEnabled",
-    "subdirExists",
-    "subdirCreate",
-    "fileExists",
-    "fileCreate",
-    "filesInsertedInRegister",
-    "filesAssignedInRegister"};
-
-  return sl.join(",").toStdString();
-}
-
-std::string DirectoryStats::toCsv() const
-{
-  QStringList oss;
-
-  auto s = [](auto ns) {
-    return ns.count() / 1000.0 / 1000.0 / 1000.0;
-  };
-
-  oss
-    << QString::number(s(dirTimes))
-    << QString::number(s(fileTimes))
-    << QString::number(s(sortTimes))
-
-    << QString::number(s(subdirLookupTimes))
-    << QString::number(s(addDirectoryTimes))
-
-    << QString::number(s(filesLookupTimes))
-    << QString::number(s(addFileTimes))
-    << QString::number(s(addOriginToFileTimes))
-    << QString::number(s(addFileToOriginTimes))
-    << QString::number(s(addFileToRegisterTimes))
-
-    << QString::number(originExists)
-    << QString::number(originCreate)
-    << QString::number(originsNeededEnabled)
-
-    << QString::number(subdirExists)
-    << QString::number(subdirCreate)
-
-    << QString::number(fileExists)
-    << QString::number(fileCreate)
-    << QString::number(filesInsertedInRegister)
-    << QString::number(filesAssignedInRegister);
-
-  return oss.join(",").toStdString();
-}
-
-
 DirectoryEntry::DirectoryEntry(
   std::wstring name, DirectoryEntry *parent, int originID) :
     m_OriginConnection(new OriginConnection),
